+++ src/AdminCoreServiceProvider.php
@@ -8,6 +8,7 @@ use Illuminate\Support\ServiceProvider;
 use Ngos\AdminCore\Console\AdminCoreFieldCommand;
 use Ngos\AdminCore\Console\AdminCoreInstallCommand;
 use Ngos\AdminCore\Console\AdminCoreMakeCommand;
+use Ngos\AdminCore\Console\AdminCorePageCommand;
 use Ngos\AdminCore\Console\AdminCorePortalCommand;
 use Ngos\AdminCore\Console\AdminCoreReinstallCommand;
 use Ngos\AdminCore\Console\AdminCoreUninstallCommand;
@@ -34,6 +35,7 @@ class AdminCoreServiceProvider extends ServiceProvider
                 AdminCoreInstallCommand::class,
                 AdminCoreMakeCommand::class,
                 AdminCoreFieldCommand::class,
+                AdminCorePageCommand::class,
                 AdminCorePortalCommand::class,
                 AdminCoreReinstallCommand::class,
                 AdminCoreUninstallCommand::class,